Move Sentry logger to integrations
