+++ tests/conftest.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 import pytest
 from rest_framework.test import APIClient
 
-from utils.sentry import SentryLogger
+from tilavarauspalvelu.integrations.sentry import SentryLogger
 
 from tests.factories._base import FakerEN, FakerFI, FakerSV
 from tests.helpers import GraphQLClient, patch_method
