+++ tests/test_external_services/test_verkkokauppa/test_merchant_requests.py
@@ -6,6 +6,7 @@ from typing import Any
 import pytest
 from django.conf import settings
 
+from tilavarauspalvelu.integrations.sentry import SentryLogger
 from tilavarauspalvelu.integrations.verkkokauppa.merchants.exceptions import CreateMerchantError, UpdateMerchantError
 from tilavarauspalvelu.integrations.verkkokauppa.merchants.types import (
     CreateMerchantParams,
@@ -14,7 +15,6 @@ from tilavarauspalvelu.integrations.verkkokauppa.merchants.types import (
     UpdateMerchantParams,
 )
 from tilavarauspalvelu.integrations.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
-from utils.sentry import SentryLogger
 
 from tests.helpers import patch_method
 from tests.mocks import MockResponse
