+++ tests/test_external_services/test_verkkokauppa/test_payment_types.py
@@ -8,9 +8,9 @@ from typing import Any
 import pytest
 from django.conf import settings
 
+from tilavarauspalvelu.integrations.sentry import SentryLogger
 from tilavarauspalvelu.integrations.verkkokauppa.payment.exceptions import ParsePaymentError
 from tilavarauspalvelu.integrations.verkkokauppa.payment.types import Payment
-from utils.sentry import SentryLogger
 
 from tests.helpers import patch_method
