+++ tests/test_external_services/test_verkkokauppa/test_webhooks/test_order_payment_webhooks.py
@@ -7,9 +7,9 @@ from django.urls import reverse
 
 from tilavarauspalvelu.enums import OrderStatus, ReservationStateChoice
 from tilavarauspalvelu.integrations.email.main import EmailService
+from tilavarauspalvelu.integrations.sentry import SentryLogger
 from tilavarauspalvelu.integrations.verkkokauppa.payment.exceptions import GetPaymentError
 from tilavarauspalvelu.integrations.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
-from utils.sentry import SentryLogger
 
 from tests.factories import PaymentOrderFactory, ReservationFactory
 from tests.helpers import patch_method
