+++ tests/test_external_services/test_verkkokauppa/test_webhooks/test_order_refund_webhooks.py
@@ -6,9 +6,9 @@ import pytest
 from django.urls import reverse
 
 from tilavarauspalvelu.enums import OrderStatus
+from tilavarauspalvelu.integrations.sentry import SentryLogger
 from tilavarauspalvelu.integrations.verkkokauppa.payment.exceptions import GetRefundStatusError
 from tilavarauspalvelu.integrations.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
-from utils.sentry import SentryLogger
 
 from tests.factories import PaymentOrderFactory
 from tests.helpers import patch_method
