+++ tests/test_graphql_api/test_helsinki_profile/test_query.py
@@ -6,7 +6,7 @@ import pytest
 
 from tilavarauspalvelu.integrations.helauth.clients import HelsinkiProfileClient
 from tilavarauspalvelu.integrations.helauth.typing import ADLoginAMR, LoginMethod, ProfileLoginAMR
-from utils.sentry import SentryLogger
+from tilavarauspalvelu.integrations.sentry import SentryLogger
 
 from tests.factories import ApplicationFactory, ReservationFactory, UserFactory
 from tests.factories.helsinki_profile import MyProfileDataFactory
