+++ tests/test_graphql_api/test_reservation/test_confirm.py
@@ -7,10 +7,10 @@ from django.test import override_settings
 from graphene_django_extensions.testing import build_mutation
 
 from tilavarauspalvelu.enums import OrderStatus, PaymentType, ReservationNotification, ReservationStateChoice
+from tilavarauspalvelu.integrations.sentry import SentryLogger
 from tilavarauspalvelu.integrations.verkkokauppa.order.exceptions import CreateOrderError
 from tilavarauspalvelu.integrations.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
 from tilavarauspalvelu.models import PaymentOrder
-from utils.sentry import SentryLogger
 
 from tests.factories import (
     OrderFactory,
