+++ tests/test_utils/test_generate_reservations_from_allocations.py
@@ -18,6 +18,7 @@ from tilavarauspalvelu.enums import (
 )
 from tilavarauspalvelu.integrations.opening_hours.hauki_api_client import HaukiAPIClient
 from tilavarauspalvelu.integrations.opening_hours.hauki_api_types import HaukiAPIDatePeriod
+from tilavarauspalvelu.integrations.sentry import SentryLogger
 from tilavarauspalvelu.models import (
     AffectingTimeSpan,
     RecurringReservation,
@@ -26,7 +27,6 @@ from tilavarauspalvelu.models import (
 )
 from tilavarauspalvelu.tasks import generate_reservation_series_from_allocations
 from utils.date_utils import DEFAULT_TIMEZONE, combine, local_date, local_datetime, local_iso_format
-from utils.sentry import SentryLogger
 
 from tests.factories import AllocatedTimeSlotFactory, ReservationFactory
 from tests.helpers import patch_method
