+++ tilavarauspalvelu/admin/request_log/admin.py
@@ -13,9 +13,9 @@ from rangefilter.filters import DateTimeRangeFilterBuilder, NumericRangeFilterBu
 
 from tilavarauspalvelu.admin.helpers import ImmutableModelAdmin
 from tilavarauspalvelu.admin.sql_log.admin import SQLLogAdminInline
+from tilavarauspalvelu.integrations.sentry import SentryLogger
 from tilavarauspalvelu.models.request_log.model import RequestLog
 from tilavarauspalvelu.services.csv_export import SQLLogCSVExporter
-from utils.sentry import SentryLogger
 
 if TYPE_CHECKING:
     from django.http import FileResponse
